Extract builder out to its own source file
